Refactor runWithRetry out of container-loader into driver-utils
