+++ packages/loader/driver-utils/package.json
@@ -57,6 +57,7 @@
   "dependencies": {
     "@fluidframework/common-definitions": "^0.20.0-0",
     "@fluidframework/common-utils": "^0.30.0-0",
+    "@fluidframework/container-definitions": "^0.40.0",
     "@fluidframework/core-interfaces": "^0.40.0",
     "@fluidframework/driver-definitions": "^0.40.0",
     "@fluidframework/gitresources": "^0.1025.0-0",
