+++ packages/loader/driver-utils/src/index.ts
@@ -15,3 +15,4 @@ export * from "./fluidResolvedUrl";
 export * from "./summaryForCreateNew";
 export * from "./parallelRequests";
 export * from "./networkUtils";
+export * from "./runWithRetry";
