+++ packages/loader/driver-utils/src/runWithRetry.ts
@@ -1,0 +1,78 @@
+/*!
+ * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+import { v4 as uuid } from "uuid";
+import { ITelemetryLogger } from "@fluidframework/common-definitions";
+import { performance } from "@fluidframework/common-utils";
+import { ICriticalContainerError, IDeltaDelayInfo } from "@fluidframework/container-definitions";
+import { canRetryOnError, getRetryDelayFromError } from "./network";
+
+const delay = async (timeMs: number): Promise<void> =>
+    new Promise((resolve) => setTimeout(() => resolve(), timeMs));
+
+export async function runWithRetry<T>(
+    api: () => Promise<T>,
+    fetchCallName: string,
+    deltaDelayInfo: IDeltaDelayInfo,
+    convertError: (error: any) => ICriticalContainerError,
+    logger: ITelemetryLogger,
+    shouldRetry?: () => { retry: boolean, error: any | undefined},
+): Promise<T> {
+    let result: T | undefined;
+    let success = false;
+    let retryAfterMs = 1000; // has to be positive!
+    let numRetries = 0;
+    const startTime = performance.now();
+    let lastError: any;
+    let id: string | undefined;
+    do {
+        try {
+            result = await api();
+            if (id !== undefined) {
+                deltaDelayInfo.refreshDelayInfo(id);
+            }
+            success = true;
+        } catch (err) {
+            if (shouldRetry !== undefined) {
+                const res = shouldRetry();
+                if (res.retry === false) {
+                    if (res.error !== undefined) {
+                        throw res.error;
+                    }
+                    throw err;
+                }
+            }
+            // If it is not retriable, then just throw the error.
+            if (!canRetryOnError(err)) {
+                logger.sendErrorEvent({
+                    eventName: fetchCallName,
+                    retry: numRetries,
+                    duration: performance.now() - startTime,
+                }, err);
+                throw err;
+            }
+            numRetries++;
+            lastError = err;
+            // If the error is throttling error, then wait for the specified time before retrying.
+            // If the waitTime is not specified, then we start with retrying immediately to max of 8s.
+            retryAfterMs = getRetryDelayFromError(err) ?? Math.min(retryAfterMs * 2, 8000);
+            if (id === undefined) {
+                id = uuid();
+            }
+            deltaDelayInfo.emitDelayInfo(id, retryAfterMs, convertError(err));
+            await delay(retryAfterMs);
+        }
+    } while (!success);
+    if (numRetries > 0) {
+        logger.sendTelemetryEvent({
+            eventName: fetchCallName,
+            retry: numRetries,
+            duration: performance.now() - startTime,
+        },
+        lastError);
+    }
+    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+    return result!;
+}
